pyproject.toml
[tool.poetry]
name = "oaipmh"
version = "0.0.1"
description = "runs OAI-PMH service."
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "656c8e841a2c610019f19fff8cc8d13d0983e377"}

flask = "*"

opentelemetry-api = "^1.13.0"
opentelemetry-exporter-gcp-trace = "^1.3.0"
opentelemetry-instrumentation-flask = "^0.34b0"
opentelemetry-instrumentation-requests = "^0.34b0"
opentelemetry-propagator-gcp = "^1.3.0"
opentelemetry-sdk = "^1.13.0"

[tool.poetry.group.dev.dependencies]
mypy = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"